requirements_no_psycopg.txt
# =====================================================
# SARAMEDICO BACKEND - PYTHON DEPENDENCIES
# =====================================================

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis & Caching
redis[hiredis]==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.0
pyotp==2.9.0
qrcode[pil]==7.4.2

# Validation & Config
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# MinIO (S3-compatible)
minio==7.2.3

# Celery & Background Tasks
celery==5.3.6
kombu==5.3.5

# Email
aiosmtplib==3.0.1
jinja2==3.1.3

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Monitoring & Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
ruff==0.1.14
isort==5.13.2

# Utilities
python-dotenv==1.0.0
boto3==1.34.0
pypdf==4.0.1
pillow==10.2.0
